requirements.txt
fastapi
uvicorn
python-multipart
python-docx
PyPDF2
langchain
openai
python-jose
passlib
sqlalchemy
pydantic
python-dotenv
flask
flask-cors
pytest
requests
Werkzeug